# pizza-order/src/main/resources/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8083",
  "basePath": "/",
  "tags": [
    {
      "name": "basic-error-controller",
      "description": "Basic Error Controller"
    },
    {
      "name": "order-rest",
      "description": "Order Rest"
    }
  ],
  "paths": {
    "/billing/createOrder": {
      "post": {
        "tags": [
          "order-rest"
        ],
        "summary": "createOrder",
        "operationId": "createOrderUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "orderEntityRequest",
            "description": "order",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderEntityRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/billing/deleteOrder": {
      "post": {
        "tags": [
          "order-rest"
        ],
        "summary": "deleteOrder",
        "operationId": "deleteOrderUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "orderId",
            "description": "orderId",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderStatusRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/billing/orderStatus": {
      "post": {
        "tags": [
          "order-rest"
        ],
        "summary": "getStatus",
        "operationId": "getStatusUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
          	"in": "body",
            "name": "orderId",
            "description": "orderId",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderStatusRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/billing/submitOrder": {
      "post": {
        "tags": [
          "order-rest"
        ],
        "summary": "submitOrder",
        "operationId": "submitOrderUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "orderId",
            "description": "orderId",
            "required": true,
            "schema": {
              "$ref": "#/definitions/OrderStatusRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
    "OrderEntityRequest": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "integer",
          "format": "int32"
        },
        "location": {
          "type": "string"
        },
        "menuId": {
          "type": "integer",
          "format": "int64"
        },
        "orderId": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "type": "string",
          "enum": [
            "ORDERED",
            "DELIVERY",
            "DELIVERED",
            "DELETED"
          ]
        }
      },
      "title": "OrderEntity"
    },
    "OrderStatusRequest":{
	      "type": "object",
	      "properties": {
	        "orderId": {
	          "type": "integer",
	          "format": "int64"
        	}
	    }
	},
    "View": {
      "type": "object",
      "properties": {
        "contentType": {
          "type": "string"
        }
      },
      "title": "View"
    }
  }
}